requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.6
Django==3.2.8
djangorestframework==3.12.4
dnspython==1.16.0
idna==3.2
mypy==0.910
mypy-extensions==0.4.3
psycopg2==2.9.1
pymongo==3.12.0
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
